\name{S_XX_study1}
\alias{S_XX_study1}
\docType{data}

\title{
Correlations between 10 SNPs corresponding to the population underlying study 1.
}

\description{
Data frame containing correlations between SNPs estimated from a reference
database matching the study 1 population, e.g., the 1000Genomes.
Here, [10 SNPs x 10 SNPs].
}


\value{Test data}


\format{Data frame}


\source{
Part of the simulated toy data set.
}


\keyword{datasets}
