\name{S_XY_full_study2}
\alias{S_XY_full_study2}
\docType{data}

\title{
Univariate summary statistics of 10 traits across 1000 SNPs (study 2).
}

\description{
Data frame containing univariate summary statistics (regression coefficients and
standard errors) of study 2 for 1000 SNPs and 10 traits. It will be used for
estimating phenotypic correlation structure \code{S_YY} of study 2.
}


\value{Test data}


\format{Data frame}


\source{
Part of the simulated toy data set.
}


\keyword{datasets}
