\name{S_XY_study1}
\alias{S_XY_study1}
\docType{data}

\title{
Univariate summary statistics of 10 traits across 10 SNPs (study 1).
}

\description{
Data frame containing univariate summary statistics (regression coefficients and
standard errors) of study 1 corresponding to the variables to be included in the
association analysis: 10 SNPs and 10 traits.
}


\value{Test data}


\format{Data frame}


\source{
Part of the simulated toy data set.
}


\keyword{datasets}
